Managed to get javafx working but had to make it work on the sample file
will have to delete the sample file
