+++ src/main/java/module-info.java
@@ -1,0 +1,8 @@
+module cash.sample {
+    requires javafx.controls;
+    requires javafx.fxml;
+
+
+    opens cash.sample to javafx.fxml;
+    exports cash.sample;
+}
