--- untitled/.idea/libraries/lib.xml
@@ -1,15 +1,0 @@
-<component name="libraryTable">
-  <library name="lib">
-    <CLASSES>
-      <root url="file://$PROJECT_DIR$/../javafx-sdk-21.0.3/lib" />
-      <root url="file://$PROJECT_DIR$/../javafx-sdk-21.0.3/lib" />
-    </CLASSES>
-    <JAVADOC />
-    <NATIVE>
-      <root url="file://$PROJECT_DIR$/../javafx-sdk-21.0.3/lib" />
-    </NATIVE>
-    <SOURCES />
-    <jarDirectory url="file://$PROJECT_DIR$/../javafx-sdk-21.0.3/lib" recursive="false" />
-    <jarDirectory url="file://$PROJECT_DIR$/../javafx-sdk-21.0.3/lib" recursive="false" />
-  </library>
-</component>
