--- untitled/project/test/Test/Test/Login/Pass.txt
@@ -1,2 +1,0 @@
-Jhon
-123
