--- untitled/project/test/Test/Test/Login/src/Main.java
@@ -1,66 +1,0 @@
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Scanner;
-import java.io.*;
-import java.util.List;
-
-
-//TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
-// click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
-public class Main {
-    Scanner input = new Scanner(System.in);
-    /**
-     * below is the database for all user ID's
-     */
-
- 
-
-
-
-
-
-
-
-
-
-
-
-
-    public static void main(String[] args) {
-        User jhon = new User("Jhon","123");
-        User lewis = new User("Lewis", "321");
-
-        HashMap<String, User> users = new HashMap<String, User>();
-
-        users.put("jhon", jhon);
-        users.put("lewis", lewis);
-
-        System.out.println(users);
-
-
-
-
-
-
-
-
-
-        }
-
-void login(){
-
-}
-
-}
-
-
-
-
-
-
-
-
-
