--- untitled/project/test/Test/Test/Login/src/User.java
@@ -1,28 +1,0 @@
-public class User {
-    private String user_name;
-    private String password;
-
-
-    public String getUser_name() {
-        return user_name;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setUser_name(String user_name) {
-        this.user_name = user_name;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public User(String user_name, String password) {
-        this.user_name = user_name;
-        this.password = password;
-    }
-
-
-}
