--- untitled/src/main/java/module-info.java
@@ -1,8 +1,0 @@
-module org.example.demo {
-    requires javafx.controls;
-    requires javafx.fxml;
-
-
-    opens org.example.demo to javafx.fxml;
-    exports org.example.demo;
-}
